# tools/docs/openapi.yaml
components:
    schemas:
        ErrorResponse:
            properties:
                error:
                    properties:
                        code:
                            description: A specific error code identifier (e.g., 'invalid_parameter_value', 'unauthorized')
                            type: string
                        message:
                            description: A human-readable error message
                            type: string
                        param:
                            description: The parameter that caused the error, if applicable
                            nullable: true
                            type: string
                        type:
                            description: The error type category (e.g., 'invalid_request_error', 'authentication_error')
                            type: string
                    required:
                        - message
                        - type
                        - code
                    type: object
            required:
                - error
            type: object
...
        array_planstore_Plan:
            items:
                $ref: '#/components/schemas/planstore_Plan'
            type: array
...
        planapi_activeResponse:
            properties:
                plan:
                    $ref: '#/components/schemas/planstore_Plan'
                steps:
                    items:
                        $ref: '#/components/schemas/planstore_PlanStep'
                    type: array
            required:
                - plan
                - steps
            type: object
        planapi_cleanResponse:
            properties:
                removed:
                    type: integer
            required:
                - removed
            type: object
        planapi_compileRequest:
            properties:
                chain_id:
                    type: string
                executor_chain_id:
                    type: string
                markdown:
                    type: string
                summarizer_chain_id:
                    type: string
                write_path:
                    type: string
            required:
                - markdown
                - executor_chain_id
                - summarizer_chain_id
                - chain_id
            type: object
        planapi_compileResponse:
            properties:
                chain:
                    $ref: '#/components/schemas/taskengine_TaskChainDefinition'
                goal:
                    type: string
                path:
                    type: string
                steps:
                    items:
                        type: string
                    type: array
            required:
                - goal
                - steps
                - chain
            type: object
        planapi_markdownResponse:
            properties:
                markdown:
                    type: string
            required:
                - markdown
            type: object
        planapi_newPlanRequest:
            properties:
                goal:
                    type: string
                planner_chain_id:
                    type: string
            required:
                - goal
                - planner_chain_id
            type: object
        planapi_newPlanResponse:
            properties:
                markdown:
                    type: string
                plan:
                    $ref: '#/components/schemas/planstore_Plan'
                steps:
                    items:
                        $ref: '#/components/schemas/planstore_PlanStep'
                    type: array
            required:
                - plan
                - steps
                - markdown
            type: object
        planapi_nextStepRequest:
            properties:
                executor_chain_id:
                    type: string
                summarizer_chain_id:
                    type: string
                with_auto:
                    type: boolean
                with_shell:
                    type: boolean
            required:
                - executor_chain_id
                - summarizer_chain_id
                - with_shell
                - with_auto
            type: object
        planapi_nextStepResponse:
            properties:
                markdown:
                    type: string
                result:
                    type: string
            required:
                - result
                - markdown
            type: object
        planapi_replanRequest:
            properties:
                planner_chain_id:
                    type: string
            required:
                - planner_chain_id
            type: object
        planapi_replanResponse:
            properties:
                markdown:
                    type: string
                steps:
                    items:
                        $ref: '#/components/schemas/planstore_PlanStep'
                    type: array
            required:
                - steps
                - markdown
            type: object
        planapi_runCompiledRequest:
            properties:
                chain_id:
                    type: string
                executor_chain_id:
                    type: string
                summarizer_chain_id:
                    type: string
                write_path:
                    type: string
            required:
                - executor_chain_id
                - summarizer_chain_id
                - chain_id
            type: object
        planapi_runCompiledResponse:
            properties:
                chain:
                    $ref: '#/components/schemas/taskengine_TaskChainDefinition'
                goal:
                    type: string
                output:
                    type: object
                output_type:
                    type: string
                path:
                    type: string
                state:
                    items:
                        $ref: '#/components/schemas/taskengine_CapturedStateUnit'
                    type: array
                steps:
                    items:
                        type: string
                    type: array
            required:
                - goal
                - steps
                - output
                - output_type
                - state
            type: object
        planstore_Plan:
            properties:
                compile_executor_chain_id:
                    description: CompileExecutorChainID is the executor chain ID used when compiling (for invalidation).
                    type: string
                compiled_chain_id:
                    description: CompiledChainID is the TaskChainDefinition.ID used when compiling (for invalidation).
                    type: string
                compiled_chain_json:
                    description: CompiledChainJSON is the full [plancompile.Compile] output JSON for incremental Next (optional).
                    type: string
                created_at:
                    format: date-time
                    type: string
                goal:
                    type: string
                id:
                    type: string
                name:
                    type: string
                session_id:
                    type: string
                status:
                    $ref: '#/components/schemas/planstore_PlanStatus'
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - name
                - goal
                - status
                - session_id
                - created_at
                - updated_at
            type: object
        planstore_PlanStatus:
            type: string
...
    securitySchemes:
        X-API-Key:
            in: header
            name: X-API-Key
            type: apiKey
info:
    title: contenox – LLM Backend Management API
    version: v0.6.9
openapi: 3.1.0
paths:
...
    /plans:
        get:
            description: Lists all plans.
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/array_planstore_Plan'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Lists all plans.
        post:
            description: |-
                Creates a new plan from a free-text goal.
                The planner_chain_id must reference an existing TaskChainDefinition.
                The chain is called with the goal text; it must return a JSON array of step strings.
                The new plan becomes the active plan.
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/planapi_newPlanRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_newPlanResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Creates a new plan from a free-text goal.
    /plans/{name}:
        delete:
            description: Permanently deletes a plan by name.
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                type: string
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Permanently deletes a plan by name.
        parameters:
            - description: The plan name to delete.
              in: path
              name: name
              required: true
              schema:
                type: string
    /plans/{name}/activate:
        parameters:
            - description: The plan name to activate.
              in: path
              name: name
              required: true
              schema:
                type: string
        put:
            description: Switches the active plan to the named plan (archives the previous active).
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                type: string
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Switches the active plan to the named plan (archives the previous active).
    /plans/active:
        get:
            description: Returns the active plan and all its steps.
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_activeResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Returns the active plan and all its steps.
    /plans/active/next:
        post:
            description: |-
                Executes the next pending step of the active plan.
                executor_chain_id must reference a TaskChainDefinition that accepts a step
                description string and returns the execution result.
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/planapi_nextStepRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_nextStepResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Executes the next pending step of the active plan.
    /plans/active/replan:
        post:
            description: |-
                Replaces remaining pending steps with a freshly generated plan.
                Completed steps are preserved; the planner is called with a recap of the
                original goal plus the completed steps to produce the new remaining steps.
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/planapi_replanRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_replanResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Replaces remaining pending steps with a freshly generated plan.
    /plans/active/run-compiled:
        post:
            description: |-
                Compiles the active plan (same markdown as Show) and executes the resulting linear chain.
                executor_chain_id and chain_id are required. Optional write_path persists the compiled chain JSON.
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/planapi_runCompiledRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_runCompiledResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Compiles the active plan (same markdown as Show) and executes the resulting linear chain.
    /plans/active/steps/{ordinal}/retry:
        parameters:
            - description: The 1-based step ordinal.
              in: path
              name: ordinal
              required: true
              schema:
                type: string
        post:
            description: Resets a step to pending so it will be retried on the next Next call.
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_markdownResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Resets a step to pending so it will be retried on the next Next call.
    /plans/active/steps/{ordinal}/skip:
        parameters:
            - description: The 1-based step ordinal.
              in: path
              name: ordinal
              required: true
              schema:
                type: string
        post:
            description: Marks a step as intentionally skipped.
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_markdownResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Marks a step as intentionally skipped.
    /plans/clean:
        post:
            description: Removes all completed or archived plans.
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_cleanResponse'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Removes all completed or archived plans.
    /plans/compile:
        post:
            description: |-
                Compiles plan Markdown into a linear TaskChainDefinition (one executor-chain copy per step).
                markdown must match the shape produced by planservice renderMarkdown. executor_chain_id must
                reference an existing task chain. Optional write_path persists the JSON via the chain VFS.
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/planapi_compileRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/planapi_compileResponse'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Default error response
            summary: Compiles plan Markdown into a linear TaskChainDefinition (one executor-chain copy per step).
...
security:
    - X-API-Key: []